Added ReactRouter to the project
