+++ frontend/package.json
@@ -19,7 +19,8 @@
     "form-data": "^4.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-redux": "^8.0.7"
+    "react-redux": "^8.0.7",
+    "react-router-dom": "^6.12.1"
   },
   "devDependencies": {
     "@types/react": "^18.0.37",
