+++ frontend/src/main.tsx
@@ -1,16 +1,19 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { Provider } from 'react-redux'
+import { BrowserRouter } from 'react-router-dom';
 
-import App from './App.tsx'
 import { store } from './store'
-
 import './styles/index.css'
 
+import App from './App.tsx'
+
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
-    <React.StrictMode>
-      <Provider store={store}>
+  <React.StrictMode>
+    <Provider store={store}>
+      <BrowserRouter>
         <App />
-      </Provider>
-    </React.StrictMode>,
+      </BrowserRouter>
+    </Provider>
+  </React.StrictMode>,
 );
