+++ frontend/src/pages/Workspace.tsx
@@ -1,0 +1,10 @@
+// import { Routes, Route, Navigate } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
+
+export default function Workspace() {
+    return (
+        <div>
+            <Outlet />
+        </div>
+    )
+};
